Modify and add migration

+++ database/migrations/2020_12_01_112829_recreate_working_days_table.php
@@ -13,14 +13,17 @@ class RecreateWorkingDaysTable extends Migration
      */
     public function up()
     {
-        Schema::create('working_days', function (Blueprint $table) {
-            $table->increments('id');
-            $table->unsignedInteger('employee_id')->nullable();
-            $table->foreign('employee_id')->references('id')->on('employees')->onDelete('no action');
-            $table->date('working_on_day')->nullable();
-            $table->time('checkin_time')->nullable();
-            $table->time('checkout_time')->nullable();
-        });
+        Schema::dropIfExists('working_days');
+        if(!Schema::hasTable('working_days')){
+            Schema::create('working_days', function (Blueprint $table) {
+                $table->increments('id');
+                $table->unsignedInteger('employee_id')->nullable();
+                $table->foreign('employee_id')->references('id')->on('employees')->onDelete('no action');
+                $table->date('working_on_day')->nullable();
+                $table->time('checkin_time')->nullable();
+                $table->time('checkout_time')->nullable();
+            });
+        }
     }
 
     /**
@@ -30,6 +33,6 @@ class RecreateWorkingDaysTable extends Migration
      */
     public function down()
     {
-        Schema::dropIfExists('working_days');
+
     }
 }
